Moved configuration check into its own class
